Use variadic templates to simplify function calling.
