+++ include/camp/detail/functionimpl.hpp
@@ -35,10 +35,9 @@
 
 
 #include <camp/function.hpp>
-#include <camp/detail/callhelper.hpp>
 #include <vector>
 #include <string>
-
+#include <utility>
 
 namespace camp {
 namespace detail {
@@ -73,6 +72,42 @@ inline typename std::remove_reference<T>::type convertArg(const Args& args, std:
 }
 
 
+template <typename R, typename C>
+struct CallHelper
+{
+    template<typename F, typename... A, std::size_t... Is>
+    static Value call(F func, C obj, const Args& args, const std::string& name, camp::util::index_sequence<Is...>)
+    {
+        return func(obj, convertArg<A>(args,Is,name)...);
+    }
+
+    template <typename F, typename... A>
+    static Value call(F func, C obj, const Args& args, const std::string& name)
+    {
+        return call<F, A...>(func, obj, args, name, camp::util::make_index_sequence<sizeof...(A)>());
+    }
+};
+
+/*
+ * Specialization of CallHelper for functions returning void
+ */
+template <typename C>
+struct CallHelper<void, C>
+{
+    template<typename F, typename... A, std::size_t... Is>
+    static Value call(F func, C obj, const Args& args, const std::string& name, camp::util::index_sequence<Is...>)
+    {
+        return func(obj, convertArg<A>(args,Is,name)...), Value::nothing;
+    }
+    
+    template <typename F, typename... A>
+    static Value call(F func, C obj, const Args& args, const std::string& name)
+    {
+        return call<F, A...>(func, obj, args, name, camp::util::make_index_sequence<sizeof...(A)>());
+    }
+};
+
+    
 /**
  * \class camp::detail::FunctionImpl
  *
@@ -84,249 +119,41 @@ inline typename std::remove_reference<T>::type convertArg(const Args& args, std:
 
 template <typename F1, typename F2 = void> class FunctionImpl;
 
+    
 /*
- * Specialization of FunctionImpl for functions taking no argument
+ * Specialization of FunctionImpl for all functions.
  */
-template <typename C, typename R>
-class FunctionImpl<R (C)> : public Function
+template <typename C, typename R, typename ...A>
+class FunctionImpl<R (C, A...)> : public Function
 {
 public:
 
     /**
      * \brief Constructor
      */
-    FunctionImpl(const std::string& name, std::function<R (C)> function)
-        : Function(name, mapType<R>())
-        , m_function(function)
+    FunctionImpl(const std::string& name, std::function<R (C, A...)> function)
+        :   Function(name, mapType<R>(), std::vector<Type> { mapType<A>()... })
+        ,   m_function(function)
     {
     }
 
 protected:
-
-    /**
-     * \see Function::execute
-     */
-    virtual Value execute(const UserObject& object, const Args&) const override
-    {
-        return CallHelper<R, C>::call(m_function, object.get<C>());
-    }
-
-private:
-
-    std::function<R (C)> m_function; ///< Object containing the actual function to call
-};
-
-/*
- * Specialization of FunctionImpl for functions taking 1 argument
- */
-template <typename C, typename R, typename A0>
-class FunctionImpl<R (C, A0)> : public Function
-{
-public:
-
-    /**
-     * \brief Constructor
-     */
-    FunctionImpl(const std::string& name, std::function<R (C, A0)> function)
-    : Function(name, mapType<R>(), std::vector<Type> { mapType<A0>() })
-        , m_function(function)
-    {
-    }
-
-protected:
-
+    
     /**
      * \see Function::execute
      */
     virtual Value execute(const UserObject& object, const Args& args) const override
     {
-        return CallHelper<R, C>::call(m_function, object.get<C>()
-                                      , convertArg<A0>(args, 0, name()));
+        return CallHelper<R, C>::template call<decltype(m_function), A...>(m_function, object.get<C>(), args, name());
     }
 
 private:
 
-    std::function<R (C, A0)> m_function; ///< Object containing the actual function to call
+    std::function<R (C, A...)> m_function; ///< Object containing the actual function to call
 };
 
 /*
- * Specialization of FunctionImpl for functions taking 2 arguments
- */
-template <typename C, typename R, typename A0, typename A1>
-class FunctionImpl<R (C, A0, A1)> : public Function
-{
-public:
-
-    /**
-     * \brief Constructor
-     */
-    FunctionImpl(const std::string& name, std::function<R (C, A0, A1)> function)
-    : Function(name, mapType<R>(), std::vector<Type> {mapType<A0>(), mapType<A1>()})
-        , m_function(function)
-    {
-    }
-
-protected:
-
-    /**
-     * \see Function::execute
-     */
-    virtual Value execute(const UserObject& object, const Args& args) const override
-    {
-        return CallHelper<R, C>::call(m_function, object.get<C>()
-                                      , convertArg<A0>(args, 0, name())
-                                      , convertArg<A1>(args, 1, name()));
-    }
-
-private:
-
-    std::function<R (C, A0, A1)> m_function; ///< Object containing the actual function to call
-};
-
-/*
- * Specialization of FunctionImpl for functions taking 3 arguments
- */
-template <typename C, typename R, typename A0, typename A1, typename A2>
-class FunctionImpl<R (C, A0, A1, A2)> : public Function
-{
-public:
-
-    /**
-     * \brief Constructor
-     */
-    FunctionImpl(const std::string& name, std::function<R (C, A0, A1, A2)> function)
-    : Function(name, mapType<R>(), std::vector<Type>{mapType<A0>(), mapType<A1>(), mapType<A2>()})
-        , m_function(function)
-    {
-    }
-
-protected:
-
-    /**
-     * \see Function::execute
-     */
-    virtual Value execute(const UserObject& object, const Args& args) const override
-    {
-        return CallHelper<R, C>::call(m_function, object.get<C>()
-                                      , convertArg<A0>(args, 0, name())
-                                      , convertArg<A1>(args, 1, name())
-                                      , convertArg<A2>(args, 2, name()));
-    }
-
-private:
-
-    std::function<R (C, A0, A1, A2)> m_function; ///< Object containing the actual function to call
-};
-
-/*
- * Specialization of FunctionImpl for functions taking 4 arguments
- */
-template <typename C, typename R, typename A0, typename A1, typename A2, typename A3>
-class FunctionImpl<R (C, A0, A1, A2, A3)> : public Function
-{
-public:
-
-    /**
-     * \brief Constructor
-     */
-    FunctionImpl(const std::string& name, std::function<R (C, A0, A1, A2, A3)> function)
-    : Function(name, mapType<R>(), std::vector<Type> {mapType<A0>(), mapType<A1>(), mapType<A2>(), mapType<A3>()})
-        , m_function(function)
-    {
-    }
-
-protected:
-
-    /**
-     * \see Function::execute
-     */
-    virtual Value execute(const UserObject& object, const Args& args) const override
-    {
-        return CallHelper<R, C>::call(m_function, object.get<C>()
-                                      , convertArg<A0>(args, 0, name())
-                                      , convertArg<A1>(args, 1, name())
-                                      , convertArg<A2>(args, 2, name())
-                                      , convertArg<A3>(args, 3, name()));
-    }
-
-private:
-
-    std::function<R (C, A0, A1, A2, A3)> m_function; ///< Object containing the actual function to call
-};
-
-/*
- * Specialization of FunctionImpl for functions taking 5 arguments
- */
-template <typename C, typename R, typename A0, typename A1, typename A2, typename A3, typename A4>
-class FunctionImpl<R (C, A0, A1, A2, A3, A4)> : public Function
-{
-public:
-
-    /**
-     * \brief Constructor
-     */
-    FunctionImpl(const std::string& name, std::function<R (C, A0, A1, A2, A3, A4)> function)
-    : Function(name, mapType<R>(), std::vector<Type> {mapType<A0>(), mapType<A1>(), mapType<A2>(), mapType<A3>(), mapType<A4>()})
-        , m_function(function)
-    {
-    }
-
-protected:
-
-    /**
-     * \see Function::execute
-     */
-    virtual Value execute(const UserObject& object, const Args& args) const override
-    {
-        return CallHelper<R, C>::call(m_function, object.get<C>()
-                                      , convertArg<A0>(args, 0, name())
-                                      , convertArg<A1>(args, 1, name())
-                                      , convertArg<A2>(args, 2, name())
-                                      , convertArg<A3>(args, 3, name())
-                                      , convertArg<A4>(args, 4, name()));
-    }
-
-private:
-
-    std::function<R (C, A0, A1, A2, A3, A4)> m_function; ///< Object containing the actual function to call
-};
-
-
-/*
- * Specialization of FunctionImpl for composed functions taking no argument
- */
-template <typename C, typename N, typename M, typename R>
-class FunctionImpl<R (N), M (C)> : public Function
-{
-public:
-
-    /**
-     * \brief Constructor
-     */
-    template <typename F1, typename F2>
-    FunctionImpl(const std::string& name, F1 function, F2 accessor)
-        : Function(name, mapType<R>())
-        , m_function(std::bind(function, std::bind(accessor, _1)))
-    {
-    }
-
-protected:
-
-    /**
-     * \see Function::execute
-     */
-    virtual Value execute(const UserObject& object, const Args&) const override
-    {
-        return CallHelper<R, C>::call(m_function, object.get<C>());
-    }
-
-private:
-
-    std::function<R (C)> m_function; ///< Object containing the actual function to call
-};
-
-/*
- * Specialization of FunctionImpl for composed functions taking 1 argument
+ * Specialization of FunctionImpl for composed functions.
  */
 template <typename C, typename N, typename M, typename R, typename A0>
 class FunctionImpl<R (N, A0), M (C)> : public Function
